refactor: move memory utilities to src/memory.ts
