+++ packages/core/src/index.browser.ts
@@ -24,6 +24,7 @@ export * from './actions';
 export * from './database';
 export * from './entities';
 export * from './logger';
+export * from './memory';
 export * from './prompts';
 export * from './roles';
 export * from './runtime';
