+++ packages/core/src/memory.ts
@@ -1,0 +1,113 @@
+import {
+  type Memory,
+  type MemoryMetadata,
+  type MessageMemory,
+  type DocumentMetadata,
+  type FragmentMetadata,
+  type MessageMetadata,
+  type DescriptionMetadata,
+  type CustomMetadata,
+  MemoryType,
+  type Content,
+  type UUID,
+} from './types';
+
+/**
+ * Factory function to create a new message memory with proper defaults
+ */
+export function createMessageMemory(params: {
+  id?: UUID;
+  entityId: UUID;
+  agentId?: UUID;
+  roomId: UUID;
+  content: Content & { text: string };
+  embedding?: number[];
+}): MessageMemory {
+  return {
+    ...params,
+    createdAt: Date.now(),
+    metadata: {
+      type: MemoryType.MESSAGE,
+      timestamp: Date.now(),
+      scope: params.agentId ? 'private' : 'shared',
+    },
+  };
+}
+
+/**
+ * Type guard to check if a memory metadata is a DocumentMetadata
+ * @param metadata The metadata to check
+ * @returns True if the metadata is a DocumentMetadata
+ */
+export function isDocumentMetadata(metadata: MemoryMetadata): metadata is DocumentMetadata {
+  return metadata.type === MemoryType.DOCUMENT;
+}
+
+/**
+ * Type guard to check if a memory metadata is a FragmentMetadata
+ * @param metadata The metadata to check
+ * @returns True if the metadata is a FragmentMetadata
+ */
+export function isFragmentMetadata(metadata: MemoryMetadata): metadata is FragmentMetadata {
+  return metadata.type === MemoryType.FRAGMENT;
+}
+
+/**
+ * Type guard to check if a memory metadata is a MessageMetadata
+ * @param metadata The metadata to check
+ * @returns True if the metadata is a MessageMetadata
+ */
+export function isMessageMetadata(metadata: MemoryMetadata): metadata is MessageMetadata {
+  return metadata.type === MemoryType.MESSAGE;
+}
+
+/**
+ * Type guard to check if a memory metadata is a DescriptionMetadata
+ * @param metadata The metadata to check
+ * @returns True if the metadata is a DescriptionMetadata
+ */
+export function isDescriptionMetadata(metadata: MemoryMetadata): metadata is DescriptionMetadata {
+  return metadata.type === MemoryType.DESCRIPTION;
+}
+
+/**
+ * Type guard to check if a memory metadata is a CustomMetadata
+ * @param metadata The metadata to check
+ * @returns True if the metadata is a CustomMetadata
+ */
+export function isCustomMetadata(metadata: MemoryMetadata): metadata is CustomMetadata {
+  return (
+    metadata.type !== MemoryType.DOCUMENT &&
+    metadata.type !== MemoryType.FRAGMENT &&
+    metadata.type !== MemoryType.MESSAGE &&
+    metadata.type !== MemoryType.DESCRIPTION
+  );
+}
+
+/**
+ * Memory type guard for document memories
+ */
+export function isDocumentMemory(
+  memory: Memory
+): memory is Memory & { metadata: DocumentMetadata } {
+  return memory.metadata?.type === MemoryType.DOCUMENT;
+}
+
+/**
+ * Memory type guard for fragment memories
+ */
+export function isFragmentMemory(
+  memory: Memory
+): memory is Memory & { metadata: FragmentMetadata } {
+  return memory.metadata?.type === MemoryType.FRAGMENT;
+}
+
+/**
+ * Safely access the text content of a memory
+ * @param memory The memory to extract text from
+ * @param defaultValue Optional default value if no text is found
+ * @returns The text content or default value
+ */
+export function getMemoryText(memory: Memory, defaultValue = ''): string {
+  return memory.content.text ?? defaultValue;
+}
